Allow importing client without cross-fetch
