+++ src/GraphQLClient.ts
@@ -1,0 +1,94 @@
+import { ClientError, GraphQLError, Headers as HttpHeaders, Options, Variables } from './types'
+import getResult from './getResult'
+
+export default class GraphQLClient {
+  private url: string
+  private options: Options
+
+  constructor(url: string, options?: Options) {
+    this.url = url
+    this.options = options || {}
+  }
+
+  async rawRequest<T extends any>(
+    query: string,
+    variables?: Variables,
+  ): Promise<{ data?: T, extensions?: any, headers: Headers, status: number, errors?: GraphQLError[] }> {
+    const { headers, ...others } = this.options
+
+    const body = JSON.stringify({
+      query,
+      variables: variables ? variables : undefined,
+    })
+
+    const response = await fetch(this.url, {
+      method: 'POST',
+      headers: Object.assign({ 'Content-Type': 'application/json' }, headers),
+      body,
+      ...others,
+    })
+
+    const result = await getResult(response)
+
+    if (response.ok && !result.errors && result.data) {
+      const { headers, status } = response
+      return { ...result, headers, status }
+    } else {
+      const errorResult =
+        typeof result === 'string' ? { error: result } : result
+      throw new ClientError(
+        { ...errorResult, status: response.status, headers: response.headers },
+        { query, variables },
+      )
+    }
+  }
+
+  async request<T extends any>(
+    query: string,
+    variables?: Variables,
+  ): Promise<T> {
+    const { headers, ...others } = this.options
+
+    const body = JSON.stringify({
+      query,
+      variables: variables ? variables : undefined,
+    })
+
+    const response = await fetch(this.url, {
+      method: 'POST',
+      headers: Object.assign({ 'Content-Type': 'application/json' }, headers),
+      body,
+      ...others,
+    })
+
+    const result = await getResult(response)
+
+    if (response.ok && !result.errors && result.data) {
+      return result.data
+    } else {
+      const errorResult =
+        typeof result === 'string' ? { error: result } : result
+      throw new ClientError(
+        { ...errorResult, status: response.status },
+        { query, variables },
+      )
+    }
+  }
+
+  setHeaders(headers: HttpHeaders): GraphQLClient {
+    this.options.headers = headers
+
+    return this
+  }
+
+  setHeader(key: string, value: string): GraphQLClient {
+    const { headers } = this.options
+
+    if (headers) {
+      headers[key] = value
+    } else {
+      this.options.headers = { [key]: value }
+    }
+    return this
+  }
+}
