+++ src/getResult.ts
@@ -1,0 +1,8 @@
+export default async function getResult(response: Response): Promise<any> {
+  const contentType = response.headers.get('Content-Type')
+  if (contentType && contentType.startsWith('application/json')) {
+    return response.json()
+  } else {
+    return response.text()
+  }
+}
